unify functions to fetch normalized implications
