+++ databases/catdat/scripts/redundancies.ts
@@ -1,9 +1,5 @@
 import { get_client } from './utils/helpers'
-import {
-	get_categories,
-	get_normalized_category_implications,
-	type NormalizedCategoryImplication,
-} from './utils/categories'
+import { get_categories } from './utils/categories'
 import {
 	get_deduced_satisfied_properties,
 	get_deduced_unsatisfied_properties,
@@ -12,8 +8,12 @@ import {
 	get_property_assignments_by_deduction,
 	type StructureMeta,
 } from './utils/deduction'
-import { get_functors, get_normalized_functor_implications } from './utils/functors'
+import { get_functors } from './utils/functors'
 import type { StructureType } from './config'
+import {
+	get_normalized_functor_implications,
+	NormalizedImplication,
+} from './utils/implications'
 
 const db = get_client()
 
@@ -36,11 +36,7 @@ function check_redundancies() {
 function check_redundant_property_assignments(type: StructureType) {
 	console.info(`\n--- Check redundant ${type} property assignments ---`)
 
-	// TODO: refactor this when > 2 types of structures are available
-	const implications =
-		type === 'category'
-			? get_normalized_category_implications(db)
-			: get_normalized_functor_implications(db)
+	const implications = get_normalized_functor_implications(db, type)
 
 	const structures: StructureMeta[] =
 		type === 'category' ? get_categories(db) : get_functors(db)
@@ -60,7 +56,8 @@ function check_redundant_property_assignments(type: StructureType) {
 			assignments[structure.id].satisfied.non_deduced,
 			implications,
 			ignore_dict[structure.id],
-			structure.associated_satisfied_properties,
+			structure.source_props,
+			structure.target_props,
 		)
 
 		if (redundant_satisfied_property) {
@@ -81,7 +78,8 @@ function check_redundant_property_assignments(type: StructureType) {
 			assignments[structure.id].unsatisfied.non_deduced,
 			implications,
 			ignore_dict[structure.id],
-			structure.associated_satisfied_properties,
+			structure.source_props,
+			structure.target_props,
 		)
 
 		if (redundant_unsatisfied_property) {
@@ -112,9 +110,10 @@ function check_redundant_property_assignments(type: StructureType) {
  */
 function get_redundant_satisfied_property(
 	satisfied_properties: Set<string>,
-	implications: NormalizedCategoryImplication[],
+	implications: NormalizedImplication[],
 	ignored: Set<string> = new Set(),
-	associated_satisfied_properties?: Record<string, Set<string>>,
+	source_props?: Set<string>,
+	target_props?: Set<string>,
 ) {
 	for (const p of [...satisfied_properties]) {
 		if (ignored.has(p)) continue
@@ -124,7 +123,8 @@ function get_redundant_satisfied_property(
 			implications,
 			{ stop_when_found: p },
 			'category',
-			associated_satisfied_properties,
+			source_props,
+			target_props,
 		)
 		if (deduced_satisfied_properties.has(p)) return p
 		satisfied_properties.add(p)
@@ -141,9 +141,10 @@ function get_redundant_satisfied_property(
 function get_redundant_unsatisfied_property(
 	satisfied_properties: Set<string>,
 	unsatisfied_properties: Set<string>,
-	implications: NormalizedCategoryImplication[],
+	implications: NormalizedImplication[],
 	ignored: Set<string> = new Set(),
-	associated_satisfied_properties?: Record<string, Set<string>>,
+	source_props?: Set<string>,
+	target_props?: Set<string>,
 ) {
 	for (const p of [...unsatisfied_properties]) {
 		if (ignored.has(p)) continue
@@ -154,7 +155,8 @@ function get_redundant_unsatisfied_property(
 			implications,
 			{ stop_when_found: p },
 			'category',
-			associated_satisfied_properties,
+			source_props,
+			target_props,
 		)
 		if (deduced_unsatisfied_properties.has(p)) return p
 		unsatisfied_properties.add(p)
